Refactor images
