+++ src/assets/images/nav-bar-icon.png
